requirement.txt
pandas>=0.23.4
Flask>=1.1.2
opencv-python>=4.5.5.64
keras>=2.2.0
tensorflow>=1.9.0
numpy>=1.14.0
Pillow>=5.2.0
gdown>=3.10.1
mysql-connector-python
google-cloud-vision